% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{buildComparison}
\alias{buildComparison}
\title{Build T vs C comparison object}
\usage{
buildComparison(targetId, comparatorId)
}
\arguments{
\item{targetId}{The Id of the target cohort}

\item{comparatorId}{The Id of the comparator cohort}
}
\value{
An object containing the targetId and comparatorId
}
\description{
Logically groups targets and comparator Ids
}
