% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateCovariates}
\alias{generateCovariates}
\title{(NOT WORKING YET) Wrapper for FeatureExtraction to generate covariates of cohort}
\usage{
generateCovariates(connectionDetails, cdmDatabaseSchema, cdmVersion, cohortId)
}
\arguments{
\item{connectionDetails}{The connection details of the database, as generated using \code{createConnectionDetails}}

\item{cdmDatabaseSchema}{The fully qualified name of the database schema holding the CDM data}

\item{cdmVersion}{The version of the CDM database}

\item{cohortId}{The Id of the cohort of interest}
}
\description{
Wrapper for FeatureExtraction to generate covariates of a T vs C pair
}
