% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateSqlCovariates}
\alias{generateSqlCovariates}
\title{Wrapper for Sql script to generate covariates of cohort}
\usage{
generateSqlCovariates(connectionDetails, repoConnectionDetails,
  cdmDatabaseSchema, ohdsiRepositorySchema, scratchDatabaseSchema, tablePrefix,
  webApiPrefix, useHttps, cdmVersion, comparisons)
}
\arguments{
\item{connectionDetails}{The connection details of the database, as generated using \code{createConnectionDetails}}

\item{repoConnectionDetails}{The connection details of the OHDSI Repo database, as generated using \code{createConnectionDetails}}

\item{cdmDatabaseSchema}{The fully qualified name of the database schema holding the CDM data}

\item{ohdsiRepositorySchema}{The fully qualified name of the database schema holding the OHDSI Repo}

\item{scratchDatabaseSchema}{The fully qualified name of the scratch database schema}

\item{tablePrefix}{The prefix of all scratch tables}

\item{webApiPrefix}{The URL name of the WebAPI instance}

\item{useHttps}{Should SSL be used?}

\item{cdmVersion}{The version of the CDM database}

\item{comparisons}{List of cohort comparisons, each of type \code{buildComparison}}
}
\description{
Wrapper for Sql script to generate covariates
}
