% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortComparison.R
\name{getChartData}
\alias{getChartData}
\title{Get Chart Data}
\usage{
getChartData(connectionDetails, tablePrefix, cdmDatabaseSchema,
  resultsDatabaseSchema, scratchDatabaseSchema, cohortDf)
}
\arguments{
\item{connectionDetails}{A connectionDetails object created using \code{createConnectionDetails}}

\item{tablePrefix}{The prefix of the scratch tables}

\item{cdmDatabaseSchema}{The fully qualified schema name of the CDM database schema}

\item{resultsDatabaseSchema}{The fully qualified schema name of the results database schema}

\item{scratchDatabaseSchema}{The fully qualified schema name of the scratch database schema}

\item{cohortDf}{A data frame with COMPARISON_ID, TARGET_ID, COMPARATOR_ID,
and COMPARISON_NAME}
}
\description{
Get Chart Data
}
\details{
This function can be used to save chart data for
         use in the plotly chart
}
