% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webApi.R
\name{getCohortDefName}
\alias{getCohortDefName}
\title{Get Cohort Definition Name}
\usage{
getCohortDefName(webApiPrefix, cohortDefinitionId, useHttps = FALSE)
}
\arguments{
\item{webApiPrefix}{The URL for WebAPI}

\item{cohortDefinitionId}{The concept set id in Atlas}

\item{useHttps}{Should SSL be used?}
}
\value{
The verbatim name of the cohort
}
\details{
Obtains the name of a cohort
}
\author{
Ajit Londhe, Jamie Weaver
}
