% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webApi.R
\name{getFormattedCohortName}
\alias{getFormattedCohortName}
\title{Get Formatted Cohort Name}
\usage{
getFormattedCohortName(webApiPrefix, cohortDefinitionId, useHttps = FALSE)
}
\arguments{
\item{webApiPrefix}{The URL prefix of the WebAPI instance}

\item{cohortDefinitionId}{The cohort definition Id}

\item{useHttps}{Should SSL be used?}
}
\value{
A string with the formatted cohort name
}
\details{
Gets a formatted name for a cohort
}
\author{
Ajit Londhe, Jamie Weaver
}
