% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPlotCovariates}
\alias{getPlotCovariates}
\title{(NOT WORKING YET) Get Plot Covariates}
\usage{
getPlotCovariates(covariateData)
}
\arguments{
\item{covariateData}{A data frame with covariate data generated by \code{generateCovariates}}
}
\value{
A subsetted data frame ready for visualization
}
\description{
Subsets the full covariate data RDS file
}
