% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortComparison.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{Creates a data frame summarizing the cohort comparison}
\usage{
getSummaryTable(dataFolder, comparisonsPath, baseUrl)
}
\arguments{
\item{dataFolder}{The path to the folder containing data files produced by \code{\link{getChartData}}}

\item{comparisonsPath}{The path to the CSV file with the comparisons}

\item{baseUrl}{The base URL of the WebAPI instance}
}
\description{
Creates a data frame summarizing the cohort comparison
}
