% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortComparison.R
\name{importFeaturesFromWebApi}
\alias{importFeaturesFromWebApi}
\title{Import Features From WebAPI}
\usage{
importFeaturesFromWebApi(baseUrl, cohortId, sourceKey)
}
\arguments{
\item{baseUrl}{The URL of the WebAPI endpoint}

\item{cohortId}{The cohort definition id in Atlas}

\item{sourceKey}{The source key for the CDM as defined in Atlas}
}
\value{
A list of data frames (one for distributed features, one for prevalence features),
                    with name and df as attributes
}
\description{
Import Features From WebAPI
}
