% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{launchXyPlot}
\alias{launchXyPlot}
\title{Launch the XY Plot Shiny app}
\usage{
launchXyPlot(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the XY Plot Shiny app
}
\details{
Launches a Shiny app that allows the viewing of covariate SMDs between T and C in an X-Y plot.
}
