% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortComparison.R
\name{getChartData}
\alias{getChartData}
\title{Get Chart Data}
\usage{
getChartData(connectionDetails, vocabDatabaseSchema, resultsDatabaseSchema,
  generationId, targetCohortId, comparatorCohortId,
  outputFolder = "output")
}
\arguments{
\item{connectionDetails}{A connectionDetails object created using \code{createConnectionDetails}}

\item{vocabDatabaseSchema}{The fully qualified schema name of the CDM database schema}

\item{resultsDatabaseSchema}{The fully qualified schema name of the results database schema}

\item{generationId}{The ID of the Cohort Characterization generation from Atlas}

\item{targetCohortId}{The ID of the target cohort as defined in Atlas}

\item{comparatorCohortId}{The ID of the comparator cohort as defined in Atlas}

\item{outputFolder}{A folder to store the chart data RDS files}
}
\description{
Get Chart Data
}
\details{
This function can be used to save chart data for
         use in the plotly chart
}
